requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
websockets
aiofiles
requests
beautifulsoup4
reportlab
cachetools
slowapi
playwright
aiohttp
asyncpg
sqlalchemy[asyncio]
alembic
passlib
bcrypt==3.2.0
python-jose[cryptography]
pydantic[email]
pydantic-settings
openai
pinecone-client
langchain
langchain-openai
langchain-pinecone
aiosqlite
python-dotenv
loguru


email-validator


stripe


sse-starlette


watchdog

